# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: 🎣 굴비잇기 API 서버
  description: >
    굴비잇기 프로젝트의 API 서버입니다.

    ## 📣 응답

    `200`번대 응답은 성공했을 때의 응답으로, 엔드포인트에 따라 바디가 있을 수도,
    없을 수도 있습니다.
    <br><br>
    `400`번대 응답은 `422 Unprocessable Entity`를 제외하고는 모든 엔드포인트에
    적용되는 공통 에러 응답입니다. <br>따라서 중복을 막기 위해 `/health` 엔드포인트를
    제외하고는 응답란에서 제외시켰습니다.


    ## ⛔️ 에러코드

    에러코드는 4자리 숫자로 구성되며, 앞 2자리는 그룹을, 뒷 2자리는 그룹 내 구분자
    역할을 합니다. <br>
    예를 들어, `1002`는 10번 그룹 2번째 에러를 의미합니다.<br><br>

    자세한 에러코드 및 메시지 정보는 **아래 링크👇**에서 확인할 수 있습니다. <br>
    [에러코드 목록](https://github.com/Central-MakeUs/goolbitg-server/blob/spec/v1/docs/errors.md)
  termsOfService: http://localhost:8080/terms-of-service
  version: 1.0.3
servers:
  - url: https://dev.goolbitg.site
    description: 개발 서버
  - url: https://api.goolbitg.site
    description: 메인 서버
  - url: http://localhost:8080
    description: 로컬호스트 서버
paths:
...
  /api/v1/images:
    post:
      summary: 이미지 업로드
      operationId: uploadImage
      description: 이미지를 업로드합니다.
      tags:
        - ETC
      requestBody:
        required: true
        description: 업로드할 이미지입니다.
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
              required:
                - image
            encoding:
              image:
                contentType: image/*
      responses:
        '201':
          description: 업로드 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageUploadResponse'
        '422':
          description: 업로드 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                허용되지 않은 이미지 타입:
                  $ref: '#/components/examples/Error_ImageTypeNotAllowed'
                너무 큰 이미지:
                  $ref: '#/components/examples/Error_ImageSizeTooHuge'
...
  /api/v1/admin/sendChallengeNotice:
    post:
      summary: 미완료 챌린지 안내 알림 전송
      operationId: sendChallengeNotice
      description: > 
        미완료 챌린지를 안내하는 알림을 모든 유저에게 전송합니다.
        <br><br>
        admin password를 제공해야 합니다.
      tags:
        - ETC
      parameters:
        - name: password
          description: admin password
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 알림 전송 성공
...
  /api/v1/buyOrNots:
    get:
      summary: 살까말까 포스트 목록 가져오기
      operationId: getBuyOrNots
      description: >
        살까말까 포스트 목록을 최신순으로 가져옵니다.<br><br>
        목록에는 개인화가 적용될 수 있습니다.
      tags:
        - BuyOrNot
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - name: created
          description: 내가 작성한 포스트만 가져올지 여부입니다.
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 살까말까 포스트 최신순 목록 가져오기
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBuyOrNotDto'
    post:
      summary: 살까말까 포스트 등록하기
      operationId: createBuyOrNot
      description: 살까말까 포스트를 등록합니다.
      tags:
        - BuyOrNot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyOrNotDto'
            examples:
              일반:
                $ref: '#/components/examples/BuyOrNot_Form_Normal'
      responses:
        '201':
          description: 살까말까 포스트 등록 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyOrNotDto'
        '422':
          description: 살까말까 등록 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                등록 한도 초과:
                  $ref: '#/components/examples/Error_PostLimitExceed'
  /api/v1/buyOrNots/{postId}:
    put:
      summary: 살까말까 포스트 수정하기
      operationId: updateBuyOrNot
      description: 살까말까 포스트를 수정합니다.
      tags:
        - BuyOrNot
      parameters:
        - $ref: '#/components/parameters/postIdParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyOrNotDto'
            examples:
              일반:
                $ref: '#/components/examples/BuyOrNot_Form_Modify'
      responses:
        '200':
          description: 살까말까 수정 완료. 수정된 정보를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyOrNotDto'
        '422':
          description: 살까말까 수정 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                존재하지 않는 포스트:
                  $ref: '#/components/examples/Error_PostNotExist'
    get:
      summary: 특정 살까말까 포스트 정보 확인
      operationId: getBuyOrNot
      description: 특정 살까말까 포스트의 정보를 확인합니다.
      tags:
        - BuyOrNot
      parameters:
        - $ref: '#/components/parameters/postIdParam'
      responses:
        '200':
          description: 특정 살까말까 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyOrNotDto'
        '422':
          description: 살까말까 수정 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                존재하지 않는 포스트:
                  $ref: '#/components/examples/Error_PostNotExist'
    delete:
      summary: 살까말까 포스트 삭제
      operationId: deleteBuyOrNot
      description: 특정 살까말까 포스트를 삭제합니다.
      tags:
        - BuyOrNot
      parameters:
        - $ref: '#/components/parameters/postIdParam'
      responses:
        '200':
          description: 살까말까 제거 완료
        '422':
          description: 살까말까 수정 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                존재하지 않는 포스트:
                  $ref: '#/components/examples/Error_PostNotExist'
  /api/v1/buyOrNots/{postId}/report:
    post:
      summary: 살까말까 포스트 신고
      operationId: buyOrNotReport
      description: 살까말까 포스트를 신고합니다.
      tags:
        - BuyOrNot
      parameters:
        - $ref: '#/components/parameters/postIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: 신고 사유
              required: 
                - reason
      responses:
        '200':
          description: 살까말까 신고 완료
        '422':
          description: 살까말까 신고 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                존재하지 않는 포스트:
                  $ref: '#/components/examples/Error_PostNotExist'
                이미 신고한 포스트:
                  $ref: '#/components/examples/Error_PostAlreadyReported'
  /api/v1/buyOrNots/{postId}/vote:
    post:
      summary: 살까말까 투표
      operationId: voteBuyOrNot
      description: 특정 살까말까 포스트에 투표합니다.
      tags:
        - BuyOrNot
      parameters:
        - $ref: '#/components/parameters/postIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyOrNotVoteDto'
      responses:
        '200':
          description: >
            살까말까 투표 완료<br><br>
            투표 완료후 투표수를 반환합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyOrNotVoteChangeDto'
        '422':
          description: 살까말까 투표 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                내가 작성한 포스트:
                  $ref: '#/components/examples/Error_CannotVoteMyPost'
...
components:
  schemas:
    ImageUploadResponse:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: 업로드한 이미지 url
...
    PaginatedBuyOrNotDto:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseDto'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/BuyOrNotDto'
...
    BuyOrNotDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 살까말까 등록 ID
        writerId:
          type: string
          description: 작성자 ID
        productName:
          type: string
          description: 상품 이름
        productPrice:
          type: integer
          minimum: 1
          maximum: 10000000
          description: 상품 금액
        productImageUrl:
          type: string
          format: uri
          description: 상품 사진 URL
        goodReason:
          type: string
          minimum: 2
          maximum: 25 # TODO: Determine min, max
          description: 사야하는 이유
        badReason:
          type: string
          minimum: 2
          maximum: 25 # TODO: Determine min, max
          description: 사지 말아야 하는 이유
        goodVoteCount:
          type: integer
          description: 살까 투표 개수
        badVoteCount:
          type: integer
          description: 말까 투표 개수
      required:
        - productName
        - productPrice
        - productImageUrl
        - goodReason
        - badReason
      example:
        id: 0
        writerId: id0001
        productName: Tecket 후드티
        productPrice: 97000
        productImageUrl: http://localhost:8080/image/bn1
        goodReason: 후드티 안산지 벌써 2년 다됨
        badReason: 집에 후드티만 10장 있긴함
        goodVoteCount: 10
        badVoteCount: 3
    BuyOrNotVoteDto:
      type: object
      properties:
        vote:
          $ref: '#/components/schemas/BuyOrNotVoteType'
          description: 사는 것에 투표할지 여부
    BuyOrNotVoteChangeDto:
      type: object
      properties:
        goodVoteCount:
          type: integer
          description: 살까 투표 개수
        badVoteCount:
          type: integer
          description: 말까 투표 개수
    BuyOrNotVoteType:
      type: string
      enum:
        - GOOD
        - BAD
...
    ErrorDto:
      type: object
      properties:
        code:
          type: integer
          minimum: 1000
          maximum: 9000
          description: 오류 코드
        message:
          type: string
          minimum: 1
          maximum: 100
          description: 오류 메세지
      required:
        - code
        - message
      example:
        code: 1000
        message: 알 수 없는 서버 오류
    PaginatedResponseDto:
      type: object
      properties:
        totalSize:
          type: integer
          description: 총 아이템 수
          example: 31
        totalPages:
          type: integer
          description: 총 페이지 수
          example: 4
        size:
          type: integer
          default: 10
          description: 아이템 개수
        page:
          type: integer
          minimum: 0
          default: 0
          description: 페이지 번호 (0부터 시작)
        items: {}
      required:
        - totalSize
        - totalPages
        - size
        - page
        - items
...
  examples:
...
    Error_PostLimitExceed:
      value:
        code: 5001
        message: 포스트 등록 한도를 초과하였습니다.
    Error_PostNotExist:
      value:
        code: 5002
        message: 포스트가 존재하지 않습니다. (post id)
    Error_CannotVoteMyPost:
      value:
        code: 5003
        message: 내가 작성한 포스트는 투표할 수 없습니다. (post id)
    Error_PostAlreadyReported:
      value:
        code: 5004
        message: 이미 신고한 포스트입니다. (post id)
...
    Error_ImageTypeNotAllowed:
      value:
        code: 7001
        message: 허용되지 않는 이미지 타입입니다. (type)
    Error_ImageSizeTooHuge:
      value:
        code: 7002
        message: 이미지는 최대 10MB 까지만 업로드할 수 있습니다.
...
    BuyOrNot_Form_Normal:
      value:
        productName: "Tecket 후드티"
        productPrice: 97000
        productImageUrl: "http://localhost:8080/image/bn1"
        goodReason: "후드티 안산지 벌써 2년 다됨"
        badReason: "집에 후드티만 10장 있긴함"
    BuyOrNot_Form_Modify:
      value:
        productName: "Tecket 후드티"
        productPrice: 90000
        productImageUrl: "http://localhost:8080/image/bn2"
        goodReason: "후드티 안산지 벌써 2년 다됨"
        badReason: "집에 후드티 12장 있음 ㅋㅋ"
  parameters:
    pageParam:
      name: page
      description: 페이지 번호입니다. 0부터 시작합니다.
      in: query
      schema:
        type: integer
        default: 0
    sizeParam:
      name: size
      description: 페이지 사이즈입니다. 기본 10입니다.
      in: query
      schema:
        type: integer
        default: 10
...
    postIdParam:
      name: postId 
      description: 살까말까 포스트 아이디입니다.
      in: path
      required: true
      schema:
        type: integer
        format: int64
        default: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...